+++ src/main/java/wad/seoul_nolgoat/web/review/dto/response/ReviewDetailsForStoreDto.java
@@ -8,7 +8,7 @@ import lombok.RequiredArgsConstructor;
 public class ReviewDetailsForStoreDto {
 
     private final Long id;
-    private final double grade;
+    private final int grade;
     private final String content;
     private final String imageUrl;
     private final Long userId;
